Move `_cache_transform` to its own file in `qml.workflow` (#6624)
**Context:**

Introduces a new file `_cache_transform.py` in `qml.workflow` to house
the private `_cache_transform` helper. Previously it lived in
`execution.py` despite having its own unit test file
`workflow/test_cache_transform.py`.

Also, update file name: `resolve_diff_method.py` &#8594;
`_resolve_diff_method.py` as I should have done earlier. 😄

**Benefits:** Code consistency and clean-up

---------
