+++ pennylane/workflow/_resolve_diff_method.py
