+++ pennylane/workflow/execution.py
@@ -23,8 +23,7 @@ differentiation support.
 
 import inspect
 import logging
-import warnings
-from collections.abc import Callable, MutableMapping
+from collections.abc import Callable
 from functools import partial
 from typing import Literal, Optional, Union, get_args
 from warnings import warn
@@ -32,10 +31,10 @@ from warnings import warn
 from cachetools import Cache, LRUCache
 
 import pennylane as qml
-from pennylane.tape import QuantumScript, QuantumScriptBatch
-from pennylane.transforms import transform
-from pennylane.typing import Result, ResultBatch
+from pennylane.tape import QuantumScriptBatch
+from pennylane.typing import ResultBatch
 
+from ._cache_transform import _cache_transform
 from .jacobian_products import DeviceDerivatives, DeviceJacobianProducts, TransformJacobianProducts
 
 logger = logging.getLogger(__name__)
@@ -93,16 +92,6 @@ INTERFACE_MAP = dict(zip(get_args(SupportedInterfaceUserInput), _mapping_output)
 
 SUPPORTED_INTERFACE_NAMES = list(INTERFACE_MAP)
 """list[str]: allowed interface strings"""
-
-_CACHED_EXECUTION_WITH_FINITE_SHOTS_WARNINGS = (
-    "Cached execution with finite shots detected!\n"
-    "Note that samples as well as all noisy quantities computed via sampling "
-    "will be identical across executions. This situation arises where tapes "
-    "are executed with identical operations, measurements, and parameters.\n"
-    "To avoid this behaviour, provide 'cache=False' to the QNode or execution "
-    "function."
-)
-"""str: warning message to display when cached execution is used with finite shots"""
 
 
 def _use_tensorflow_autograph():
@@ -215,43 +204,6 @@ def _make_inner_execute(
         return transform_post_processing(results)
 
     return inner_execute
-
-
-@transform
-def _cache_transform(tape: QuantumScript, cache: MutableMapping):
-    """Caches the result of ``tape`` using the provided ``cache``.
-
-    .. note::
-
-        This function makes use of :attr:`.QuantumTape.hash` to identify unique tapes.
-    """
-
-    def cache_hit_postprocessing(_results: ResultBatch) -> Result:
-        result = cache[tape.hash]
-        if result is not None:
-            if tape.shots and getattr(cache, "_persistent_cache", True):
-                warnings.warn(_CACHED_EXECUTION_WITH_FINITE_SHOTS_WARNINGS, UserWarning)
-            return result
-
-        raise RuntimeError(
-            "Result for tape is missing from the execution cache. "
-            "This is likely the result of a race condition."
-        )
-
-    if tape.hash in cache:
-        return [], cache_hit_postprocessing
-
-    def cache_miss_postprocessing(results: ResultBatch) -> Result:
-        result = results[0]
-        cache[tape.hash] = result
-        return result
-
-    # Adding a ``None`` entry to the cache indicates that a result will eventually be available for
-    # the tape. This assumes that post-processing functions are called in the same order in which
-    # the transforms are invoked. Otherwise, ``cache_hit_postprocessing()`` may be called before the
-    # result of the corresponding tape is placed in the cache by ``cache_miss_postprocessing()``.
-    cache[tape.hash] = None
-    return [tape], cache_miss_postprocessing
 
 
 def _get_interface_name(tapes, interface):
