+++ tests/logging/test_logging_autograd.py
@@ -80,7 +80,7 @@ class TestLogging:
                 ["Calling <construct(self=<QNode: device='<default.qubit device"],
             ),
             (
-                "pennylane.workflow.resolve_diff_method",
+                "pennylane.workflow._resolve_diff_method",
                 ["Calling <_resolve_diff_method("],
             ),
             (
